* upstream/master: (28 commits)
  [ioctl] Incorrect conversion between integer types (#3522)
  [action] fix incorrect conversion between integer types (#3545)
  [test] fix TestLoadBlockchainfromDB (#3521)
  [ioctl] correct Chinese usage message (#3510)
  fix err not hanled (#3509)
  add ReadHeaderTimeout (#3539)
  [iotcl] Reset config cmd (#3496)
  Update gosec.yaml
  Update ci.yaml
  Update analysis.yaml
  Delete gosec.yaml.bak
  Create gosec.yaml
  [config] move config.ActPool to actpool package refactor (#3514)
  Update ci.yaml
  Update analysis.yaml
  Update analysis.yaml
  [config] move config.Chain to blockchain package (#3511)
  remove circleci (#3498)
  [ioctl] Build block bucketlist command line into new ioctl (#3469)
  [config] remove  EVMNetworkID() and SetEVMNetworkID() (#3503)
  ...

+++ action/consignment_transfer.go
@@ -57,8 +57,8 @@ type (
 
 	// ConsignMsgEther is the consignment message format of Ethereum
 	ConsignMsgEther struct {
-		BucketIdx int    `json:"bucket"`
-		Nonce     int    `json:"nonce"`
+		BucketIdx uint64 `json:"bucket"`
+		Nonce     uint64 `json:"nonce"`
 		Recipient string `json:"recipient"`
 		Reclaim   string `json:"reclaim"`
 	}
@@ -148,8 +148,8 @@ func NewConsignMsg(sigType, recipient string, bucketIdx, nonce uint64) ([]byte, 
 	switch sigType {
 	case "Ethereum":
 		msg := ConsignMsgEther{
-			BucketIdx: int(bucketIdx),
-			Nonce:     int(nonce),
+			BucketIdx: bucketIdx,
+			Nonce:     nonce,
 			Recipient: recipient,
 			Reclaim:   _reclaim,
 		}
